+++ src/app/(routes)/assets/assets-data-table.tsx
@@ -24,7 +24,6 @@ import { isNil } from 'lodash'
 import { useCreateUpdateSheet } from '@/components/sheet/use-create-update-sheet'
 import useCustomToast from '@/hooks/use-custom-toast'
 import { AssetsSheet } from './assets-sheet'
-import { useParams } from 'next/navigation'
 import { EntityBox } from '@/components/entity-box'
 import { PaginationLimitField } from '@/components/form/pagination-limit-field'
 import { FormProvider, UseFormReturn } from 'react-hook-form'
