+++ src/app/(routes)/ledgers/ledgers-sheet.tsx
@@ -18,7 +18,7 @@ import { useIntl } from 'react-intl'
 import { z } from 'zod'
 import { isNil } from 'lodash'
 import { LoadingButton } from '@/components/ui/loading-button'
-import { useCreateLedger } from '@/client/ledgers'
+import { useCreateLedger, useUpdateLedger } from '@/client/ledgers'
 import { LedgerResponseDto } from '@/core/application/dto/ledger-response-dto'
 import { useOrganization } from '@/context/organization-provider/organization-provider-client'
 import useCustomToast from '@/hooks/use-custom-toast'
@@ -57,7 +57,7 @@ export const LedgersSheet = ({
   const { mutate: createLedger, isPending: createPending } = useCreateLedger({
     organizationId: currentOrganization.id!,
     onSuccess: (data: unknown) => {
-      const formData = data as { ledger: ILedgerType }
+      const formData = data as ILedgerType
       onSuccess?.()
       onOpenChange?.(false)
       showSuccess(
@@ -66,7 +66,7 @@ export const LedgersSheet = ({
             id: 'ledgers.toast.create.success',
             defaultMessage: 'Ledger {ledgerName} created successfully'
           },
-          { ledgerName: formData.ledger.name }
+          { ledgerName: formData.name }
         )
       )
     },
@@ -81,6 +81,29 @@ export const LedgersSheet = ({
     }
   })
 
+  const { mutate: updateLedger, isPending: updatePending } = useUpdateLedger({
+    organizationId: currentOrganization!.id!,
+    ledgerId: data?.id!,
+    onSuccess: () => {
+      onSuccess?.()
+      onOpenChange?.(false)
+      showSuccess(
+        intl.formatMessage({
+          id: 'ledgers.toast.update.success',
+          defaultMessage: 'Ledger changes saved successfully'
+        })
+      )
+    },
+    onError: () => {
+      showError(
+        intl.formatMessage({
+          id: 'ledgers.toast.update.error',
+          defaultMessage: 'Error updating Ledger'
+        })
+      )
+    }
+  })
+
   const form = useForm({
     resolver: zodResolver(FormSchema),
     defaultValues: Object.assign({}, defaultValues, ledger)
@@ -89,6 +112,8 @@ export const LedgersSheet = ({
   const handleSubmit = (data: FormData) => {
     if (mode === 'create') {
       createLedger(data)
+    } else if (mode === 'edit') {
+      updateLedger(data)
     }
 
     form.reset(defaultValues)
@@ -101,10 +126,10 @@ export const LedgersSheet = ({
   }, [mode])
 
   React.useEffect(() => {
-    if (!isNil(data)) {
+    if (mode === 'edit' && !isNil(data)) {
       form.reset(data, { keepDefaultValues: true })
     }
-  }, [data])
+  }, [data, mode])
 
   return (
     <Sheet onOpenChange={onOpenChange} {...others}>
@@ -127,6 +152,28 @@ export const LedgersSheet = ({
           </SheetHeader>
         )}
 
+        {mode === 'edit' && (
+          <SheetHeader>
+            <SheetTitle>
+              {intl.formatMessage(
+                {
+                  id: 'ledgers.sheet.edit.title',
+                  defaultMessage: 'Edit "{ledgerName}"'
+                },
+                {
+                  ledgerName: data?.name
+                }
+              )}
+            </SheetTitle>
+            <SheetDescription>
+              {intl.formatMessage({
+                id: 'ledgers.sheet.edit.description',
+                defaultMessage: 'View and edit ledger fields.'
+              })}
+            </SheetDescription>
+          </SheetHeader>
+        )}
+
         <Form {...form}>
           <form
             className="flex flex-grow flex-col"
@@ -134,7 +181,10 @@ export const LedgersSheet = ({
           >
             <Tabs defaultValue="details" className="mt-0">
               <TabsList className="mb-8 px-0">
-                <TabsTrigger value="details">
+                <TabsTrigger
+                  value="details"
+                  className="focus:outline-none focus:ring-0"
+                >
                   {intl.formatMessage({
                     id: 'ledgers.sheet.tabs.details',
                     defaultMessage: 'Ledger Details'
@@ -178,7 +228,7 @@ export const LedgersSheet = ({
                 type="submit"
                 disabled={!(form.formState.isDirty && form.formState.isValid)}
                 fullWidth
-                loading={createPending}
+                loading={createPending || updatePending}
               >
                 {intl.formatMessage({
                   id: 'common.save',
