+++ src/components/ui/tabs/index.tsx
@@ -39,7 +39,7 @@ const TabsTrigger = React.forwardRef<
   <TabsPrimitive.Trigger
     ref={ref}
     className={cn(
-      'inline-flex items-center justify-center whitespace-nowrap rounded-md px-4 py-[10px] text-sm font-medium text-shadcn-400 ring-offset-white transition-all hover:bg-white hover:text-shadcn-500 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-slate-950 focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 data-[state=active]:bg-accent data-[state=active]:text-accent-foreground data-[state=active]:shadow-sm',
+      '!focus-visible:outline-none !focus-visible:ring-2 !focus-visible:ring-slate-950 !focus-visible:ring-offset-2 inline-flex items-center justify-center whitespace-nowrap rounded-md px-4 py-[10px] text-sm font-medium text-shadcn-400 ring-offset-white transition-all hover:bg-white hover:text-shadcn-500 disabled:pointer-events-none disabled:opacity-50 data-[state=active]:bg-accent data-[state=active]:text-accent-foreground data-[state=active]:shadow-sm',
       className
     )}
     {...props}
