+++ src/core/application/dto/ledger-response-dto.ts
@@ -1,3 +1,4 @@
+import { AssetEntity } from '@/core/domain/entities/asset-entity'
 import { StatusDto } from './status.dto'
 
 export interface LedgerResponseDto {
@@ -9,4 +10,5 @@ export interface LedgerResponseDto {
   createdAt: Date
   updatedAt: Date
   deletedAt: Date | null
+  assets?: AssetEntity[]
 }
